updated to new health values (double instead of int)

+++ src/main/java/org/efreak/bukkitmanager/scripting/api/APIPlayer.java
@@ -14,7 +14,7 @@ public class APIPlayer {
 		return player.isOnline();
 	}
 	
-	public int getHealth() {
+	public double getHealth() {
 		return player.getHealth();
 	}
